fix: move test classes to separate folder
