+++ Tests/TestEntities/CollectionTestsClass.cs
@@ -1,0 +1,8 @@
+namespace Tests.TestEntities;
+
+public class CollectionTestsClass
+{
+    public string Name { get; set; } = null!;
+    public int Age { get; set; }
+    public List<CollectionTestsClass> Friends { get; set; } = new();
+}
