+++ Tests/TestEntities/CultureTestClass.cs
@@ -1,0 +1,8 @@
+namespace Tests.TestEntities;
+
+public class CultureTestClass
+{
+    public decimal Decimal { get; set; }
+    public DateTime Date { get; set; }
+    public double Double { get; set; }
+}
