+++ Tests/TestEntities/Departmnet.cs
@@ -1,0 +1,6 @@
+namespace Tests.TestEntities;
+
+public class Department
+{
+    public string Name { get; set; } = null!;
+}
