+++ Tests/TestEntities/Person.cs
@@ -3,7 +3,7 @@
     public class Person
     {
         public Guid Id { get; set; }
-        public string Name { get; set; }
+        public string Name { get; set; } = null!;
         public double Height { get; set; }
         public int Age { get; set; }
     }
