+++ Tests/TestEntities/TestClass.cs
@@ -3,6 +3,6 @@ namespace Tests.TestEntities;
 
 public class TestClass
 {
-    public string Property { get; set; }
-    public string Field;
+    public string Property { get; set; } = null!;
+    public string field;
 }
