+++ Tests/TestEntities/TrimTestsClass.cs
@@ -1,0 +1,8 @@
+namespace Tests.TestEntities;
+
+public class TrimTestsClass
+{
+    public string Short { get; set; } = null!;
+    public string Long { get; set; } = null!;
+    public string Description { get; set; } = null!;
+}
